+++ store/iavl/nodedb.go
@@ -3,7 +3,10 @@ package iavl
 import (
 	"bytes"
 	"container/list"
+	"encoding/binary"
 	"fmt"
+	"github.com/tendermint/go-amino"
+	"io"
 	"math"
 	"sort"
 	"sync"
@@ -612,4 +615,642 @@ func (ndb *nodeDB) String() string {
 		index++
 	})
 	return "-" + "\n" + str + "-"
+}
+
+// Node represents a node in a Tree.
+type Node struct {
+	key       []byte
+	value     []byte
+	hash      []byte
+	leftHash  []byte
+	rightHash []byte
+	version   int64
+	size      int64
+	leftNode  *Node
+	rightNode *Node
+	height    int8
+	persisted bool
+}
+
+// NewNode returns a new node from a key, value and version.
+func NewNode(key []byte, value []byte, version int64) *Node {
+	return &Node{
+		key:     key,
+		value:   value,
+		height:  0,
+		size:    1,
+		version: version,
+	}
+}
+
+// MakeNode constructs an *Node from an encoded byte slice.
+//
+// The new node doesn't have its hash saved or set. The caller must set it
+// afterwards.
+func MakeNode(buf []byte) (*Node, error) {
+
+	// Read node header (height, size, version, key).
+	height, n, cause := amino.DecodeInt8(buf)
+	if cause != nil {
+		return nil, errors.Wrap(cause, "decoding node.height")
+	}
+	buf = buf[n:]
+
+	size, n, cause := amino.DecodeVarint(buf)
+	if cause != nil {
+		return nil, errors.Wrap(cause, "decoding node.size")
+	}
+	buf = buf[n:]
+
+	ver, n, cause := amino.DecodeVarint(buf)
+	if cause != nil {
+		return nil, errors.Wrap(cause, "decoding node.version")
+	}
+	buf = buf[n:]
+
+	key, n, cause := amino.DecodeByteSlice(buf)
+	if cause != nil {
+		return nil, errors.Wrap(cause, "decoding node.key")
+	}
+	buf = buf[n:]
+
+	node := &Node{
+		height:  height,
+		size:    size,
+		version: ver,
+		key:     key,
+	}
+
+	// Read node body.
+
+	if node.isLeaf() {
+		val, _, cause := amino.DecodeByteSlice(buf)
+		if cause != nil {
+			return nil, errors.Wrap(cause, "decoding node.value")
+		}
+		node.value = val
+	} else { // Read children.
+		leftHash, n, cause := amino.DecodeByteSlice(buf)
+		if cause != nil {
+			return nil, errors.Wrap(cause, "deocding node.leftHash")
+		}
+		buf = buf[n:]
+
+		rightHash, _, cause := amino.DecodeByteSlice(buf)
+		if cause != nil {
+			return nil, errors.Wrap(cause, "decoding node.rightHash")
+		}
+		node.leftHash = leftHash
+		node.rightHash = rightHash
+	}
+	return node, nil
+}
+
+// String returns a string representation of the node.
+func (node *Node) String() string {
+	hashstr := "<no hash>"
+	if len(node.hash) > 0 {
+		hashstr = fmt.Sprintf("%X", node.hash)
+	}
+	return fmt.Sprintf("Node{%s:%s@%d %X;%X}#%s",
+		node.key,
+		node.value,
+		node.version,
+		node.leftHash, node.rightHash,
+		hashstr)
+}
+
+// clone creates a shallow copy of a node with its hash set to nil.
+func (node *Node) clone(version int64) *Node {
+	if node.isLeaf() {
+		panic("Attempt to copy a leaf node")
+	}
+	return &Node{
+		key:       node.key,
+		height:    node.height,
+		version:   version,
+		size:      node.size,
+		hash:      nil,
+		leftHash:  node.leftHash,
+		leftNode:  node.leftNode,
+		rightHash: node.rightHash,
+		rightNode: node.rightNode,
+		persisted: false,
+	}
+}
+
+func (node *Node) isLeaf() bool {
+	return node.height == 0
+}
+
+// Check if the node has a descendant with the given key.
+func (node *Node) has(t *ImmutableTree, key []byte) (has bool) {
+	if bytes.Equal(node.key, key) {
+		return true
+	}
+	if node.isLeaf() {
+		return false
+	}
+	if bytes.Compare(key, node.key) < 0 {
+		return node.getLeftNode(t).has(t, key)
+	}
+	return node.getRightNode(t).has(t, key)
+}
+
+// Get a key under the node.
+func (node *Node) get(t *ImmutableTree, key []byte) (index int64, value []byte) {
+	if node.isLeaf() {
+		switch bytes.Compare(node.key, key) {
+		case -1:
+			return 1, nil
+		case 1:
+			return 0, nil
+		default:
+			return 0, node.value
+		}
+	}
+
+	if bytes.Compare(key, node.key) < 0 {
+		return node.getLeftNode(t).get(t, key)
+	}
+	rightNode := node.getRightNode(t)
+	index, value = rightNode.get(t, key)
+	index += node.size - rightNode.size
+	return index, value
+}
+
+func (node *Node) getByIndex(t *ImmutableTree, index int64) (key []byte, value []byte) {
+	if node.isLeaf() {
+		if index == 0 {
+			return node.key, node.value
+		}
+		return nil, nil
+	}
+	// TODO: could improve this by storing the
+	// sizes as well as left/right hash.
+	leftNode := node.getLeftNode(t)
+
+	if index < leftNode.size {
+		return leftNode.getByIndex(t, index)
+	}
+	return node.getRightNode(t).getByIndex(t, index-leftNode.size)
+}
+
+// Computes the hash of the node without computing its descendants. Must be
+// called on nodes which have descendant node hashes already computed.
+func (node *Node) _hash() []byte {
+	if node.hash != nil {
+		return node.hash
+	}
+
+	h := tmhash.New()
+	buf := new(bytes.Buffer)
+	if err := node.writeHashBytes(buf); err != nil {
+		panic(err)
+	}
+	_, err := h.Write(buf.Bytes())
+	if err != nil {
+		panic(err)
+	}
+	node.hash = h.Sum(nil)
+
+	return node.hash
+}
+
+// Hash the node and its descendants recursively. This usually mutates all
+// descendant nodes. Returns the node hash and number of nodes hashed.
+func (node *Node) hashWithCount() ([]byte, int64) {
+	if node.hash != nil {
+		return node.hash, 0
+	}
+
+	h := tmhash.New()
+	buf := new(bytes.Buffer)
+	hashCount, err := node.writeHashBytesRecursively(buf)
+	if err != nil {
+		panic(err)
+	}
+	_, err = h.Write(buf.Bytes())
+	if err != nil {
+		panic(err)
+	}
+	node.hash = h.Sum(nil)
+
+	return node.hash, hashCount + 1
+}
+
+// validate validates the node contents
+func (node *Node) validate() error {
+	if node == nil {
+		return errors.New("node cannot be nil")
+	}
+	if node.key == nil {
+		return errors.New("key cannot be nil")
+	}
+	if node.version <= 0 {
+		return errors.New("version must be greater than 0")
+	}
+	if node.height < 0 {
+		return errors.New("height cannot be less than 0")
+	}
+	if node.size < 1 {
+		return errors.New("size must be at least 1")
+	}
+
+	if node.height == 0 {
+		// Leaf nodes
+		if node.value == nil {
+			return errors.New("value cannot be nil for leaf node")
+		}
+		if node.leftHash != nil || node.leftNode != nil || node.rightHash != nil || node.rightNode != nil {
+			return errors.New("leaf node cannot have children")
+		}
+		if node.size != 1 {
+			return errors.New("leaf nodes must have size 1")
+		}
+	} else {
+		// Inner nodes
+		if node.value != nil {
+			return errors.New("value must be nil for non-leaf node")
+		}
+		if node.leftHash == nil && node.rightHash == nil {
+			return errors.New("inner node must have children")
+		}
+	}
+	return nil
+}
+
+// Writes the node's hash to the given io.Writer. This function expects
+// child hashes to be already set.
+func (node *Node) writeHashBytes(w io.Writer) error {
+	err := amino.EncodeInt8(w, node.height)
+	if err != nil {
+		return errors.Wrap(err, "writing height")
+	}
+	err = amino.EncodeVarint(w, node.size)
+	if err != nil {
+		return errors.Wrap(err, "writing size")
+	}
+	err = amino.EncodeVarint(w, node.version)
+	if err != nil {
+		return errors.Wrap(err, "writing version")
+	}
+
+	// Key is not written for inner nodes, unlike writeBytes.
+
+	if node.isLeaf() {
+		err = amino.EncodeByteSlice(w, node.key)
+		if err != nil {
+			return errors.Wrap(err, "writing key")
+		}
+		// Indirection needed to provide proofs without values.
+		// (e.g. ProofLeafNode.ValueHash)
+		valueHash := tmhash.Sum(node.value)
+		err = amino.EncodeByteSlice(w, valueHash)
+		if err != nil {
+			return errors.Wrap(err, "writing value")
+		}
+	} else {
+		if node.leftHash == nil || node.rightHash == nil {
+			panic("Found an empty child hash")
+		}
+		err = amino.EncodeByteSlice(w, node.leftHash)
+		if err != nil {
+			return errors.Wrap(err, "writing left hash")
+		}
+		err = amino.EncodeByteSlice(w, node.rightHash)
+		if err != nil {
+			return errors.Wrap(err, "writing right hash")
+		}
+	}
+
+	return nil
+}
+
+// Writes the node's hash to the given io.Writer.
+// This function has the side-effect of calling hashWithCount.
+func (node *Node) writeHashBytesRecursively(w io.Writer) (hashCount int64, err error) {
+	if node.leftNode != nil {
+		leftHash, leftCount := node.leftNode.hashWithCount()
+		node.leftHash = leftHash
+		hashCount += leftCount
+	}
+	if node.rightNode != nil {
+		rightHash, rightCount := node.rightNode.hashWithCount()
+		node.rightHash = rightHash
+		hashCount += rightCount
+	}
+	err = node.writeHashBytes(w)
+
+	return
+}
+
+func (node *Node) aminoSize() int {
+	n := 1 +
+		amino.VarintSize(node.size) +
+		amino.VarintSize(node.version) +
+		amino.ByteSliceSize(node.key)
+	if node.isLeaf() {
+		n += amino.ByteSliceSize(node.value)
+	} else {
+		n += amino.ByteSliceSize(node.leftHash) +
+			amino.ByteSliceSize(node.rightHash)
+	}
+	return n
+}
+
+// Writes the node as a serialized byte slice to the supplied io.Writer.
+func (node *Node) writeBytes(w io.Writer) error {
+	cause := amino.EncodeInt8(w, node.height)
+	if cause != nil {
+		return errors.Wrap(cause, "writing height")
+	}
+	cause = amino.EncodeVarint(w, node.size)
+	if cause != nil {
+		return errors.Wrap(cause, "writing size")
+	}
+	cause = amino.EncodeVarint(w, node.version)
+	if cause != nil {
+		return errors.Wrap(cause, "writing version")
+	}
+
+	// Unlike writeHashBytes, key is written for inner nodes.
+	cause = amino.EncodeByteSlice(w, node.key)
+	if cause != nil {
+		return errors.Wrap(cause, "writing key")
+	}
+
+	if node.isLeaf() {
+		cause = amino.EncodeByteSlice(w, node.value)
+		if cause != nil {
+			return errors.Wrap(cause, "writing value")
+		}
+	} else {
+		if node.leftHash == nil {
+			panic("node.leftHash was nil in writeBytes")
+		}
+		cause = amino.EncodeByteSlice(w, node.leftHash)
+		if cause != nil {
+			return errors.Wrap(cause, "writing left hash")
+		}
+
+		if node.rightHash == nil {
+			panic("node.rightHash was nil in writeBytes")
+		}
+		cause = amino.EncodeByteSlice(w, node.rightHash)
+		if cause != nil {
+			return errors.Wrap(cause, "writing right hash")
+		}
+	}
+	return nil
+}
+
+func (node *Node) getLeftNode(t *ImmutableTree) *Node {
+	if node.leftNode != nil {
+		return node.leftNode
+	}
+	return t.ndb.GetNode(node.leftHash)
+}
+
+func (node *Node) getRightNode(t *ImmutableTree) *Node {
+	if node.rightNode != nil {
+		return node.rightNode
+	}
+	return t.ndb.GetNode(node.rightHash)
+}
+
+// NOTE: mutates height and size
+func (node *Node) calcHeightAndSize(t *ImmutableTree) {
+	node.height = maxInt8(node.getLeftNode(t).height, node.getRightNode(t).height) + 1
+	node.size = node.getLeftNode(t).size + node.getRightNode(t).size
+}
+
+func (node *Node) calcBalance(t *ImmutableTree) int {
+	return int(node.getLeftNode(t).height) - int(node.getRightNode(t).height)
+}
+
+// traverse is a wrapper over traverseInRange when we want the whole tree
+func (node *Node) traverse(t *ImmutableTree, ascending bool, cb func(*Node) bool) bool {
+	return node.traverseInRange(t, nil, nil, ascending, false, 0, false, func(node *Node, depth uint8) bool {
+		return cb(node)
+	})
+}
+
+// traversePost is a wrapper over traverseInRange when we want the whole tree post-order
+func (node *Node) traversePost(t *ImmutableTree, ascending bool, cb func(*Node) bool) bool {
+	return node.traverseInRange(t, nil, nil, ascending, false, 0, true, func(node *Node, depth uint8) bool {
+		return cb(node)
+	})
+}
+
+func (node *Node) traverseInRange(t *ImmutableTree, start, end []byte, ascending bool, inclusive bool, depth uint8, post bool, cb func(*Node, uint8) bool) bool {
+	if node == nil {
+		return false
+	}
+	afterStart := start == nil || bytes.Compare(start, node.key) < 0
+	startOrAfter := start == nil || bytes.Compare(start, node.key) <= 0
+	beforeEnd := end == nil || bytes.Compare(node.key, end) < 0
+	if inclusive {
+		beforeEnd = end == nil || bytes.Compare(node.key, end) <= 0
+	}
+
+	// Run callback per inner/leaf node.
+	stop := false
+	if !post && (!node.isLeaf() || (startOrAfter && beforeEnd)) {
+		stop = cb(node, depth)
+		if stop {
+			return stop
+		}
+	}
+
+	if !node.isLeaf() {
+		if ascending {
+			// check lower nodes, then higher
+			if afterStart {
+				stop = node.getLeftNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, post, cb)
+			}
+			if stop {
+				return stop
+			}
+			if beforeEnd {
+				stop = node.getRightNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, post, cb)
+			}
+		} else {
+			// check the higher nodes first
+			if beforeEnd {
+				stop = node.getRightNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, post, cb)
+			}
+			if stop {
+				return stop
+			}
+			if afterStart {
+				stop = node.getLeftNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, post, cb)
+			}
+		}
+	}
+	if stop {
+		return stop
+	}
+
+	if post && (!node.isLeaf() || (startOrAfter && beforeEnd)) {
+		stop = cb(node, depth)
+		if stop {
+			return stop
+		}
+	}
+
+	return stop
+}
+
+// Only used in testing...
+func (node *Node) lmd(t *ImmutableTree) *Node {
+	if node.isLeaf() {
+		return node
+	}
+	return node.getLeftNode(t).lmd(t)
+}
+
+// Provides a fixed-width lexicographically sortable []byte key format
+type KeyFormat struct {
+	prefix byte
+	layout []int
+	length int
+}
+
+// Create a []byte key format based on a single byte prefix and fixed width key segments each of whose length is
+// specified by by the corresponding element of layout.
+//
+// For example, to store keys that could index some objects by a version number and their SHA256 hash using the form:
+// 'c<version uint64><hash [32]byte>' then you would define the KeyFormat with:
+//
+//  var keyFormat = NewKeyFormat('c', 8, 32)
+//
+// Then you can create a key with:
+//
+//  func ObjectKey(version uint64, objectBytes []byte) []byte {
+//  	hasher := sha256.New()
+//  	hasher.Sum(nil)
+//  	return keyFormat.Key(version, hasher.Sum(nil))
+//  }
+func NewKeyFormat(prefix byte, layout ...int) *KeyFormat {
+	// For prefix byte
+	length := 1
+	for _, l := range layout {
+		length += l
+	}
+	return &KeyFormat{
+		prefix: prefix,
+		layout: layout,
+		length: length,
+	}
+}
+
+// Format the byte segments into the key format - will panic if the segment lengths do not match the layout.
+func (kf *KeyFormat) KeyBytes(segments ...[]byte) []byte {
+	key := make([]byte, kf.length)
+	key[0] = kf.prefix
+	n := 1
+	for i, s := range segments {
+		l := kf.layout[i]
+		if len(s) > l {
+			panic(fmt.Errorf("length of segment %X provided to KeyFormat.KeyBytes() is longer than the %d bytes "+
+				"required by layout for segment %d", s, l, i))
+		}
+		n += l
+		// Big endian so pad on left if not given the full width for this segment
+		copy(key[n-len(s):n], s)
+	}
+	return key[:n]
+}
+
+// Format the args passed into the key format - will panic if the arguments passed do not match the length
+// of the segment to which they correspond. When called with no arguments returns the raw prefix (useful as a start
+// element of the entire keys space when sorted lexicographically).
+func (kf *KeyFormat) Key(args ...interface{}) []byte {
+	if len(args) > len(kf.layout) {
+		panic(fmt.Errorf("keyFormat.Key() is provided with %d args but format only has %d segments",
+			len(args), len(kf.layout)))
+	}
+	segments := make([][]byte, len(args))
+	for i, a := range args {
+		segments[i] = format(a)
+	}
+	return kf.KeyBytes(segments...)
+}
+
+// Reads out the bytes associated with each segment of the key format from key.
+func (kf *KeyFormat) ScanBytes(key []byte) [][]byte {
+	segments := make([][]byte, len(kf.layout))
+	n := 1
+	for i, l := range kf.layout {
+		n += l
+		if n > len(key) {
+			return segments[:i]
+		}
+		segments[i] = key[n-l : n]
+	}
+	return segments
+}
+
+// Extracts the segments into the values pointed to by each of args. Each arg must be a pointer to int64, uint64, or
+// []byte, and the width of the args must match layout.
+func (kf *KeyFormat) Scan(key []byte, args ...interface{}) {
+	segments := kf.ScanBytes(key)
+	if len(args) > len(segments) {
+		panic(fmt.Errorf("keyFormat.Scan() is provided with %d args but format only has %d segments in key %X",
+			len(args), len(segments), key))
+	}
+	for i, a := range args {
+		scan(a, segments[i])
+	}
+}
+
+// Return the prefix as a string.
+func (kf *KeyFormat) Prefix() string {
+	return string([]byte{kf.prefix})
+}
+
+func scan(a interface{}, value []byte) {
+	switch v := a.(type) {
+	case *int64:
+		// Negative values will be mapped correctly when read in as uint64 and then type converted
+		*v = int64(binary.BigEndian.Uint64(value))
+	case *uint64:
+		*v = binary.BigEndian.Uint64(value)
+	case *[]byte:
+		*v = value
+	default:
+		panic(fmt.Errorf("keyFormat scan() does not support scanning value of type %T: %v", a, a))
+	}
+}
+
+func format(a interface{}) []byte {
+	switch v := a.(type) {
+	case uint64:
+		return formatUint64(v)
+	case int64:
+		return formatUint64(uint64(v))
+	// Provide formatting from int,uint as a convenience to avoid casting arguments
+	case uint:
+		return formatUint64(uint64(v))
+	case int:
+		return formatUint64(uint64(v))
+	case []byte:
+		return v
+	default:
+		panic(fmt.Errorf("keyFormat format() does not support formatting value of type %T: %v", a, a))
+	}
+}
+
+func formatUint64(v uint64) []byte {
+	bs := make([]byte, 8)
+	binary.BigEndian.PutUint64(bs, v)
+	return bs
+}
+
+func maxInt8(a, b int8) int8 {
+	if a > b {
+		return a
+	}
+	return b
 }
